requirements.txt
livekit>=1.0
livekit-agents[openai,deepgram,cartesia,silero,turndetector]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv~=1.0
openai~=1.0
mysql-connector-python~=9.0
fastapi~=0.115
uvicorn~=0.34
pydantic~=2.0
